Save Comment class out of CommentsHelper
it is too crowded.
